+++ src/main/java/com/springboot/topologie/models/data/SoftwareDAO.java
@@ -1,12 +1,15 @@
 package com.springboot.topologie.models.data;
 
-import com.springboot.topologie.models.Hardware;
+
 import com.springboot.topologie.models.Software;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import org.springframework.stereotype.Repository;
-import javax.transaction.Transactional;
+
+import java.util.List;
+import java.util.Optional;
 
 @RepositoryRestResource(collectionResourceRel="software",path="software")
-public interface SoftwareDAO extends JpaRepository<Software,Integer> {}
+public interface SoftwareDAO extends JpaRepository<Software,Long> {
+
+}
